refactor: fix checkstyle errors
