+++ src/main/java/duke/Duke.java
@@ -1,14 +1,21 @@
 package duke;
 
 import duke.command.Command;
+import duke.command.Parser;
 import duke.exception.DukeException;
 import duke.task.TaskList;
 
+/**
+ * Main class of Duke.
+ */
 public class Duke {
     private static Storage storage;
     private static TaskList tasks;
     private static Ui ui;
 
+    /**
+     * Constructs a duke chatbot.
+     */
     public Duke() {
         storage = new Storage("data", "duke.txt");
         ui = new Ui();
@@ -20,6 +27,9 @@ public class Duke {
         }
     }
 
+    /**
+     * Runs the duke chatbot until the 'bye' command is executed.
+     */
     public static void run() {
         Ui.greet(tasks.numberOfTasks());
         boolean isRunning = true;
