+++ src/main/java/duke/task/Task.java
@@ -1,7 +1,5 @@
 package duke.task;
 
-import duke.exception.DukeException;
-
 /**
  * An abstract class representing a task (which has a description, and completion status).
  */
