+++ src/test/java/TaskListTest.java
@@ -1,9 +1,10 @@
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import org.junit.jupiter.api.Test;
+
 import duke.exception.DukeException;
 import duke.task.TaskList;
 import duke.task.ToDo;
-import org.junit.jupiter.api.Test;
 
 public class TaskListTest {
     @Test
